@@ -3,21 +3,22 @@
   "version": "1.0.0",
   "description": "",
   "dependencies": {
-    "react": "~16.8",
-    "webpack": "^3.0.0"
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "webpack": "^5.1.3",
+    "webpack-cli": "4.1.0"
   },
   "devDependencies": {
     "@material-ui/core": "3.3.2",
     "@material-ui/icons": "3.0.1",
-    "babel-core": "^6.25.0",
-    "babel-loader": "^7.1.1",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-jsxz": "~1.0.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "~6.5.0",
-    "css-loader": "^0.28.11",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "reaxt-react-16": "github:Shakadak/reaxt-react-16"
+    "@babel/core": "^7.12.1",
+    "@babel/preset-env": "^7.12.1",
+    "@babel/preset-react": "^7.12.1",
+    "babel-loader": "^8.1.0",
+    "css-loader": "^5.0.0",
+    "css-minimizer-webpack-plugin": "^1.1.5",
+    "mini-css-extract-plugin": "^1.1.1",
+    "null-loader": "^4.0.1"
   },
   "author": "",
   "license": "ISC"
