@@ -17,8 +17,7 @@ defmodule Hello.Mixfile do
   ]
 
   defp deps, do: [
-    {:cowboy, "~> 1.0.0"},
-    {:plug, "~> 1.6.0"},
-    {:reaxt, github: "Shakadak/reaxt", tag: "2.1.3"}
+    {:plug_cowboy, "~> 2.5"},
+    {:reaxt, "~> 4.0"},
   ]
 end
